Development

+++ flowback/group/services/delegate.py
@@ -6,8 +6,7 @@ from flowback.comment.services import comment_create, comment_update, comment_de
 
 from flowback.common.services import get_object
 from flowback.group.models import GroupUserDelegator, GroupUserDelegatePool, GroupTags, GroupUserDelegate
-from flowback.group.selectors import group_user_permissions
-from flowback.group.services.group import group_notification
+from flowback.group.selectors.permission import group_user_permissions
 
 
 def group_user_delegate(*, user: int, group: int, delegate_pool_id: int, tags: list[int] = None) -> GroupUserDelegator:
@@ -15,8 +14,15 @@ def group_user_delegate(*, user: int, group: int, delegate_pool_id: int, tags: l
     delegator = group_user_permissions(user=user, group=group)
     delegate_pool = get_object(GroupUserDelegatePool, 'Delegate pool does not exist', id=delegate_pool_id, group=group)
 
-    if GroupUserDelegate.objects.filter(group_user=delegator).exists():
-        raise ValidationError('Delegator cannot be a delegate')
+    try:
+        delegate = GroupUserDelegate.objects.get(group_user=delegator)
+
+    except GroupUserDelegate.DoesNotExist:
+        pass
+
+    else:
+        if delegate.pool_id != delegate_pool_id:
+            raise ValidationError('Delegate cannot be a delegator beside to themselves')
 
     db_tags = GroupTags.objects.filter(id__in=tags, active=True).all()
 
@@ -32,29 +38,32 @@ def group_user_delegate(*, user: int, group: int, delegate_pool_id: int, tags: l
     if len(db_tags) < len(tags):
         raise ValidationError('Not all tags are available in the group')
 
-    delegate_rel = GroupUserDelegator(group_id=group, delegator_id=delegator.id,
-                                      delegate_pool_id=delegate_pool.id)
-    delegate_rel.full_clean()
-    delegate_rel.save()
+    delegate_rel, created = GroupUserDelegator.objects.get_or_create(group_id=group,
+                                                                     delegator_id=delegator.id,
+                                                                     delegate_pool_id=delegate_pool.id)
+
+    if not created:
+        raise ValidationError('User already delegated to this pool')
+
     delegate_rel.tags.add(*db_tags)
 
     return delegate_rel
 
 
 # TODO Likely needs an update
-def group_user_delegate_update(*, user_id: int, group_id: int, data):
+def group_user_delegate_update(*, user_id: int, group_id: int, delegate_pool_id: int, tags: list[int] = None):
+    tags = tags or []
+
     group_user = group_user_permissions(user=user_id, group=group_id)
 
-    tags = sum([x.get('tags', []) for x in data], [])
-    tags_rel = {rel['delegate_pool_id']: rel['tags'] for rel in data}
-    pools = [x.get('delegate_pool_id') for x in data]
+    tags_rel = {delegate_pool_id: tags}
+    pools = [delegate_pool_id]
 
     delegate_rel = GroupUserDelegator.objects.filter(delegator_id=group_user.id,
                                                      group_id=group_id,
                                                      delegate_pool__in=pools).all()
 
-
-    if tags == 0:
+    if len(tags) == 0:
         GroupUserDelegator.objects.filter(delegator=group_user, group_id=group_id, delegate_pool__in=pools).delete()
         return
 
@@ -63,6 +72,9 @@ def group_user_delegate_update(*, user_id: int, group_id: int, data):
 
     if len(delegate_rel) < len(pools):
         raise ValidationError('User is not delegator in all pools')
+
+    if GroupUserDelegator.objects.filter(~Q(delegate_pool_id=delegate_pool_id) & Q(tags__id__in=tags)):
+        raise ValidationError('User already delegated to same tag in another pool')
 
     TagsModel = GroupUserDelegator.tags.through
     TagsModel.objects.filter(groupuserdelegator__in=delegate_rel).delete()
@@ -95,17 +107,18 @@ def group_user_delegate_pool_create(*, user: int, group: int, blockchain_id: int
     if GroupUserDelegate.objects.filter(group=group, group_user=group_user).exists():
         raise ValidationError('User is already a delegator')
 
-    delegate_pool = GroupUserDelegatePool(group_id=group, blockchain_id=blockchain_id)
+    delegate_pool = GroupUserDelegatePool(group_id=group,
+                                          blockchain_id=blockchain_id,
+                                          related_notification_channel=group_user.group.notification_channel)
+
     delegate_pool.full_clean()
     delegate_pool.save()
     user_delegate = GroupUserDelegate(group_id=group,
                                       group_user=group_user,
                                       pool=delegate_pool)
+
     user_delegate.full_clean()
     user_delegate.save()
-
-    group_notification.create(sender_id=group, action=group_notification.Action.update, category='delegate',
-                              message=f'{group_user.user.username} is now a delegate in {group_user.group.name}')
 
     return delegate_pool
 
@@ -116,11 +129,17 @@ def group_user_delegate_pool_delete(*, user: int, group: int):
     delegate_user = get_object(GroupUserDelegate, group_user=group_user, group_id=group)
     delegate_pool = get_object(GroupUserDelegatePool, id=delegate_user.pool_id)
 
-    group_notification.create(sender_id=group, action=group_notification.Action.update, category='delegate',
-                              message=f'{group_user.user.username} has resigned from being a delegate in '
-                                      f'{group_user.group.name}')
-
     delegate_pool.delete()
+
+
+def group_user_delegate_pool_notification_subscribe(*, user: int, delegate_pool_id: int, **kwargs):
+    delegate_pool = GroupUserDelegatePool.objects.get(id=delegate_pool_id)
+    group_user = group_user_permissions(user=user, group=delegate_pool.group)
+
+    if not delegate_pool.groupuserdelegator_set.filter(delegator=group_user).exists():
+        raise ValidationError('User is not a delegate in the pool')
+
+    delegate_pool.notification_channel.subscribe(user=user, **kwargs)
 
 
 def group_delegate_pool_comment_create(*,
